it fkn blds (215b merge)

+++ app/src/pitstop/java/com/pitstop/models/TripEnd.java
@@ -13,7 +13,7 @@ public class TripEnd extends TripIndicator {
 
     private String mileage;
 
-    public TripEnd(long tripId, String rtcTime, String mileage) {
+    public TripEnd(int tripId, String rtcTime, String mileage) {
         super(tripId, rtcTime);
         this.mileage = mileage;
     }
